Create StockPricePredictionUsingKNNModel
It is a project based on Stock Price Prediction using KNN model. KNNRegressor model is used to predict values of Close pricing of a Stock. KNNClassifier model is used to classify whether to sell or buy stocks . The accuracy of model is evaluated using accuracy_score whereas root mean square error is also determined.